add: implementasi delete data

+++ characters.php
@@ -99,7 +99,7 @@ $data_mahasiswa = showData($query);
               </td>
               <td class='td-action'>
                 <a href="edit-data.php?id=<?=  $row['id'] ?>" class="btn-edit">Edit</a>
-                <a href="delete-data.php?id=<?=  $row['id'] ?>" class="btn-delete">Delete</a>
+                <a href="delete-data.php?id=<?=  $row['id'] ?>" class="btn-delete" onclick="return confirm('Are you sure?')">Delete</a>
               </td>
             </tr>
             <?php
